support generate multi letter image
