+++ draw.go
@@ -33,7 +33,7 @@ var defaultLetterColor = color.RGBA{0xf0, 0xf0, 0xf0, 0xf0}
 
 // Draw generates a new letter-avatar image of the given size using the given letter
 // with the given options. Default parameters are used if a nil *Options is passed.
-func Draw(size int, letter rune, options *Options) (image.Image, error) {
+func Draw(size int, letter []rune, options *Options) (image.Image, error) {
 	font := defaultFont
 	if options != nil && options.Font != nil {
 		font = options.Font
@@ -66,10 +66,10 @@ func Draw(size int, letter rune, options *Options) (image.Image, error) {
 	return drawAvatar(bgColor, letterColor, font, size, fontSize, letter)
 }
 
-func drawAvatar(bgColor, fgColor color.Color, font *truetype.Font, size int, fontSize float64, letter rune) (image.Image, error) {
+func drawAvatar(bgColor, fgColor color.Color, font *truetype.Font, size int, fontSize float64, letter []rune) (image.Image, error) {
 	dst := newRGBA(size, size, bgColor)
 
-	src, err := drawString(bgColor, fgColor, font, fontSize, string(letter))
+	src, err := drawString(bgColor, fgColor, font, fontSize, letter)
 	if err != nil {
 		return nil, err
 	}
@@ -80,12 +80,15 @@ func drawAvatar(bgColor, fgColor color.Color, font *truetype.Font, size int, fon
 	return dst, nil
 }
 
-func drawString(bgColor, fgColor color.Color, font *truetype.Font, fontSize float64, str string) (image.Image, error) {
+func drawString(bgColor, fgColor color.Color, font *truetype.Font, fontSize float64, letter []rune) (image.Image, error) {
 	c := freetype.NewContext()
 	c.SetDPI(72)
 
 	bb := font.Bounds(c.PointToFixed(fontSize))
 	w := bb.Max.X.Ceil() - bb.Min.X.Floor()
+	if len(letter) > 0 {
+		w = w + int(fontSize)*(len(letter)-1)
+	}
 	h := bb.Max.Y.Ceil() - bb.Min.Y.Floor()
 
 	dst := newRGBA(w, h, bgColor)
@@ -97,7 +100,7 @@ func drawString(bgColor, fgColor color.Color, font *truetype.Font, fontSize floa
 	c.SetFontSize(fontSize)
 	c.SetFont(font)
 
-	p, err := c.DrawString(str, fixed.Point26_6{X: 0, Y: bb.Max.Y})
+	p, err := c.DrawString(string(letter), fixed.Point26_6{X: 0, Y: bb.Max.Y})
 	if err != nil {
 		return nil, err
 	}
